requirements.txt
leafmap
fiona
streamlit
boto3
localtileserver
geopandas
folium
rasterio
xarray
pyproj
